Initial work on a preseed branch tool
This tool queries the charmhub using the v2 API and finds the relevant
revision for a track/risk combination.  It would then use this revision
to set another track/risk with this charm revision, thus preseeding a
branch with a track.  Will come in useful for the stable/zed branch.