preseed-branch-tool.py: Add --channel and --risk flags.
This change introduces the following new command line arguments:

* `--channel` allows the user to define which channel to look for
  revisions in.
* `--risk` allows to ask the program to list revisions in *all* channels
  for a specific risk.

Also now the program now prints in red when there is no revision for a
channel.

Usage examples:

List the revisions associated to latest/stable and 3.6/edge in the
rabbitmq-server charm.

    ./preseed-branch-tool.py -c rabbitmq-server --channel 3.6/edge

List the revisions associated to latest/stable and xena/edge for all the
charms defined in lp-builder-config

    ./preseed-branch-tool.py --channel xena/edge

List the revisions associated to aodh charm in all the channels with the
edge risk

    ./preseed-branch-tool.py --charm aodh --risk edge